backend/requirements.txt
Flask
mysql-connector-python
flask-cors
matplotlib
scikit-learn
joblib